v0.86.f: OTLog cleanup, new OTSettings and OTPaths
**** Requires GCC 4.6.0 ****

This commit depends on GCC 4.6.0 (or greater). The main reason for this is
smart pointers, (allowing making the OT much more secure and reliable).

Unfortunately, I have not been able to support older compilers than GCC
4.6  -  This is from no lack of great effort.   I was able to get it
working with GCC 4.4, however GCC 4.5 was broken.  There work-around would
have crippled much of the C++x11 features; so there would be not much
point in moving to features C++x11 anyway.

I made the choice to just make a clean break and only support GCC 4.6.
This will allow us to make use of much of the C++x11 features;
additionally remove any dependencies on BOOST (initially only for UNIX,
however in the future for Windows as-well)!!

In this commit itself we have a new OTSettings class, that makes life much
easier and cleaner than having all the settings code sit-around in OTLog.
The same thing has been done with the new OPaths class, the functionally
of this class used to be in OTLog.

I guess this will be somewhat of a special commit to be accepted by
FellowTraveller's  (When it is/if it is) as the first C++ work that I
did for OT (or any project ever) was the Settings and Paths code.  In this
commit I have finally got it to the stage where I'm "happy"÷ with it.
(Well at least it is now "good enough").

Cam.